refactor: convert images to webp
