+++ src/feature/discount-codes/constants.ts
@@ -1,5 +1,5 @@
-import imageGrind from '~/assets/img/grind-code-image.jpg';
-import imageHello from '~/assets/img/hello-code-image.jpg';
+import imageGrind from '~/assets/img/grind-code-image.webp';
+import imageHello from '~/assets/img/hello-code-image.webp';
 
 export const DISCOUNT_CODE_IMAGE_MAP = new Map<string, string>([
   ['GRIND-10', imageGrind],
