+++ src/feature/hero-banner/hero-banner.module.css
@@ -1,6 +1,6 @@
 .banner {
   background-color: var(--color-dark);
-  background-image: url('../../assets/img/banner-background.png');
+  background-image: url('../../assets/img/banner-background.webp');
   background-repeat: no-repeat;
   background-position: top center;
   background-size: auto;
